expose NewRawInputSequence to prevent costly calls to the reflect library
In most cases, users of the api will simply pass a string to the
tokenizer to get a list of tokens to pass to the model. In that case,
there is no need to call the reflect library at all, which might save
important cpu cycles and consequently improve the throughput of the
services using the tokenizer library.